+++ docs/CHANGELOG.rst
@@ -2,6 +2,14 @@ Changelog
 =========
 
 
+v1.1.0
+------
+
+Features
+  * :issue:`53`: Emit a signal when an email address is verified.
+  * :issue:`54`: Normalize email addresses.
+
+
 v1.0.0
 ------